#481 Image exporting fixed (DefineBitsJPEG2/3 erroneous header)

+++ trunk/src/com/jpexs/decompiler/flash/SWFInputStream.java
@@ -171,7 +171,7 @@ public class SWFInputStream extends InputStream {
      *
      * @param pos Number of bytes
      */
-    public void seek(int pos) throws IOException {
+    public void seek(long pos) throws IOException {
         if (is instanceof SeekableInputStream) {
             SeekableInputStream sis = (SeekableInputStream) is;
             sis.seek(pos);
